+++ tests/conftest.py
@@ -1,5 +1,7 @@
 # encoding: utf-8
 
+from datetime import date
+
 import pytest
 from mock import Mock
 
@@ -13,7 +15,7 @@ except ImportError:
     Concat = Mock()
 
 from .models import (ApplicationWithClassBasedProperties, ApplicationWithDecoratorBasedProperties,
-                     CategoryWithClassBasedProperties, CategoryWithDecoratorBasedProperties)
+                     CategoryWithClassBasedProperties, CategoryWithDecoratorBasedProperties, Version)
 
 
 @pytest.fixture
@@ -48,9 +50,14 @@ def versions(applications):
     objs = []
     for application in applications:
         objs.extend([
-            application.versions.create(major=1, minor=2, patch=3),
-            application.versions.create(major=1, minor=3, patch=0),
-            application.versions.create(major=1, minor=3, patch=1),
-            application.versions.create(major=2, minor=0, patch=0, changes='Amazing new features'),
+            application.versions.create(major=1, minor=2, patch=3, release_type=Version.BETA,
+                                        supported_from=date(2016, 7, 1), supported_until=date(2016, 12, 31)),
+            application.versions.create(major=1, minor=3, patch=0,
+                                        supported_from=date(2017, 1, 1), supported_until=date(2017, 12, 31)),
+            application.versions.create(major=1, minor=3, patch=1,
+                                        supported_from=date(2018, 1, 1), supported_until=date(2018, 12, 31)),
+            application.versions.create(major=2, minor=0, patch=0, changes='Amazing new features',
+                                        release_type=Version.ALPHA,
+                                        supported_from=date(2018, 11, 1), supported_until=date(2019, 1, 31)),
         ])
     return objs